DIVIDED THE CODE IN CLASS and LINT
