+++ HW2_OOP/src/animals/Animal.java
@@ -1,0 +1,17 @@
+package animals;
+
+import management.Supervisor;
+
+public interface Animal {
+    Integer getId();
+
+    double getHeight();
+
+    boolean canMakeSound();
+
+    void setSupervisor(Supervisor supervisor);
+
+    Supervisor getSupervisor();
+
+    String makeSound();
+}
