+++ HW2_OOP/src/animals/Cat.java
@@ -1,0 +1,12 @@
+package animals;
+
+class Cat extends AbstractAnimal {
+    public Cat(int id, double height) {
+        super(id, height);
+    }
+
+    @Override
+    public String makeSound() {
+        return "Meow";
+    }
+}
