+++ HW2_OOP/src/animals/Dog.java
@@ -1,0 +1,12 @@
+package animals;
+
+class Dog extends AbstractAnimal {
+    public Dog(int id, double height) {
+        super(id, height);
+    }
+
+    @Override
+    public String makeSound() {
+        return "Woof";
+    }
+}
