+++ HW2_OOP/src/animals/Fish.java
@@ -1,0 +1,17 @@
+package animals;
+
+class Fish extends AbstractAnimal {
+    public Fish(int id, double height) {
+        super(id, height);
+    }
+
+    @Override
+    public boolean canMakeSound() {
+        return false;
+    }
+
+    @Override
+    public String makeSound() {
+        return "";
+    }
+}
