+++ HW2_OOP/src/animals/Hippo.java
@@ -1,0 +1,12 @@
+package animals;
+
+class Hippo extends AbstractAnimal {
+    public Hippo(int id, double height) {
+        super(id, height);
+    }
+
+    @Override
+    public String makeSound() {
+        return "Grunt";
+    }
+}
