+++ HW2_OOP/src/animals/Horse.java
@@ -1,0 +1,12 @@
+package animals;
+
+class Horse extends AbstractAnimal {
+    public Horse(int id, double height) {
+        super(id, height);
+    }
+
+    @Override
+    public String makeSound() {
+        return "Neigh";
+    }
+}
