+++ HW2_OOP/src/management/Supervisor.java
@@ -1,0 +1,7 @@
+package management;
+
+public interface Supervisor {
+    Integer getId();
+
+    String getName();
+}
