+++ HW2_OOP/src/management/zooObserver.java
@@ -1,0 +1,7 @@
+package management;
+
+import animals.Animal;
+
+public interface zooObserver {
+    void onSupervisorChanged(Animal animal, Supervisor oldSupervisor, Supervisor newSupervisor);
+}
